[subtitle] header 구조 변경, 아이콘 추가
