+++ src/enum.ts
@@ -1,1 +1,7 @@
 export const NAVER_API_URL = 'https://openapi.naver.com/v1/nid/me'
+
+export const subtitles = {
+  main: '전체 메모',
+  mypage: '마이페이지',
+  etc: '기타',
+}
